Split basic app and server execution so we can just include the app without running the server
